requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.4
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
docarray==0.13.30
docker==4.2.2
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.0
gitdb==4.0.9
GitPython==3.1.27
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
jina==3.6.11
Jinja2==3.1.2
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
lz4==4.0.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rich==12.5.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.19.1
streamlit==1.10.0
streamlit-imagegrid==0.0.3
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uvicorn==0.18.2
validators==0.20.0
watchdog==2.1.9
watchfiles==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.8.1
